Make separate helper for hook and component
It also changes the parameter of useShareProvider
to an object at all instances
