+++ entry_types/scrolled/package/spec/frontend/features/scrollToSectionMessage-spec.js
@@ -1,7 +1,7 @@
 import {renderEntry, usePageObjects} from 'support/pageObjects';
+import {fakeParentWindow} from 'support';
 
-import {act} from '@testing-library/react';
-import {fakeParentWindow, tick} from 'support';
+import {asyncHandlingOf} from 'support/asyncHandlingOf';
 
 import useScrollTarget from 'frontend/useScrollTarget';
 jest.mock('frontend/useScrollTarget');
@@ -23,9 +23,8 @@ describe('SCROLL_TO_SECTION message', () => {
       }
     });
 
-    await act(async () => {
+    await asyncHandlingOf(() => {
       window.postMessage({type: 'SCROLL_TO_SECTION', payload: {index: 1}}, '*');
-      await tick(); // Make sure async handling of message is wrapped in act
     });
 
     expect(useScrollTarget.lastTarget).toBe(getSectionByPermaId(11).el);
