+++ entry_types/scrolled/package/spec/support/asyncHandlingOf/forHooks.js
