+++ entry_types/scrolled/package/spec/support/asyncHandlingOf/index.js
@@ -1,0 +1,11 @@
+import {act} from '@testing-library/react';
+import {tick} from "support";
+
+export async function asyncHandlingOf(callback) {
+  await act(async () => {
+    callback();
+    // Ensure React update triggered by async handling of message is
+    // wrapped in act
+    await tick();
+  });
+}
